fix: change function format.phone location
