+++ webapp/src/pages/User/settings/EditProfileTab.tsx
@@ -11,6 +11,7 @@ import {
   ListItemText
 } from '@mui/material';
 
+import formatPhoneNumber from 'src/services/format.phone';
 import useAuth from '../../../hooks/useAuth';
 import EditTwoToneIcon from '@mui/icons-material/EditTwoTone';
 import DoneTwoToneIcon from '@mui/icons-material/DoneTwoTone';
@@ -68,18 +69,6 @@ function EditProfileTab() {
     if ('error' in userResponse) return;
     setUser(userResponse);
   }
-
-  //Provisoire
-  function formatPhoneNumber(phoneNumberString) {
-    var cleaned = ('' + phoneNumberString).replace(/\D/g, '');
-    var match = cleaned.match(/^(\d{2})(\d{2})(\d{2})(\d{2})(\d{2})$/);
-    if (match) {
-      return match[1] + ' ' + match[2] + ' ' + match[3] + ' ' + match[4] + ' ' + match[5];
-    }
-    return null;
-  }
-
-  console.log(user);
 
   const formatedPhoneNumber = (formatPhoneNumber(user.phone));
